Move repositories interfaces from user bundle into component
